[ZK queue cleanup tool] Add arguments to define type or refstatus
Sometimes the queue got events that later should be processed by Zuul,
so we need to skip them.

+++ scripts/zk-cleanup-queue.py
@@ -16,6 +16,7 @@
 
 
 import argparse
+import json
 
 import kazoo.client
 import kazoo.exceptions
@@ -29,6 +30,13 @@ parser.add_argument('--queue', help="provide path to the queue",
                     required=True)
 parser.add_argument('--host', help="Specify zookeeper host",
                     default='zookeeper')
+parser.add_argument('--all', help="Remove all events no matter what is "
+                    "the type",
+                    action="store_true")
+parser.add_argument('--type', help="Remove specific event type"
+                    "e.g. ref-replication-scheduled or dropped-output ")
+parser.add_argument('--refstatus', help="Remove specific ref status"
+                    "eg. NON_EXISTING")
 args = parser.parse_args()
 
 client = kazoo.client.KazooClient(hosts=args.host)
@@ -38,8 +46,20 @@ events = client.get_children(args.queue)
 patches = list(map(lambda n: args.queue.rstrip("/") + "/" + n, events))
 print("Current lenght of the queue %s is %s" % (args.queue, len(patches)))
 
-for p in patches:
-    try:
-        client.delete(p)
-    except kazoo.exceptions.NoNodeError:
-        print("Can not remove %s" % p)
+if args.all:
+    for p in patches:
+        try:
+            client.delete(p)
+        except kazoo.exceptions.NoNodeError:
+            print("Can not remove %s" % p)
+elif args.type or args.refstatus:
+    for p in patches:
+        try:
+            data, _ = client.get(p)
+            jdata = json.loads(data).get('payload')
+            if (jdata.get('type', '') == args.type) or (
+                    jdata.get('refStatus', '') == args.refstatus):
+                print("Cleaning %s that contain %s" % (p, jdata))
+                client.delete(p)
+        except kazoo.exceptions.NoNodeError:
+            print("Can not remove %s" % p)
